Create Timeline
Enumeration of treeNodes is handled based on if there is a game matching the year criteria within the timeline time-frame. There is a hardcoded limit to the year dates as it proved difficult to make this a truly dyanmic range within the timef-rame that I had to work on it.